requirements.txt
aiohttp==2.3.2
async-timeout==2.0.0
bleach==2.1.1
certifi==2017.11.5
chardet==3.0.4
cssselect==1.0.1
decorator==4.1.2
entrypoints==0.2.3
flake8==3.5.0
html5lib==1.0b10
idna==2.6
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.3
jedi==0.11.0
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.4.0
lxml==4.1.1
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.8.1
multidict==3.3.2
mypy==0.540
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.1
numpy==1.13.3
pandas==0.21.0
pandocfilters==1.4.2
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.4.34
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
PyPDF2==1.26.0
pytest==3.2.3
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2017.3
pyzmq==16.0.3
qtconsole==4.3.1
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
tabula-py==1.0.0
terminado==0.6
testpath==0.3.1
tornado==4.5.2
tqdm==4.19.4
traitlets==4.3.2
typed-ast==1.1.0
Unidecode==0.4.21
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.0.6
xlrd==1.1.0
xlwt==1.3.0
yarl==0.13.0
git+https://github.com/betafcc/scrapetools
